Update: refoctoring default.vue & make AppHeader.vue component
